name override for pow
