+++ pennylane/ops/op_math/pow2.py
@@ -15,7 +15,7 @@
 """Defines the base class for the power of operators."""
 
 from functools import reduce
-from typing import Union
+from typing import Union, override
 
 from scipy.linalg import fractional_matrix_power
 
@@ -106,15 +106,23 @@ class Pow2(SymbolicOp2):
         )
 
     @property
+    @override
+    def name(self) -> str:
+        return f"Pow({self.base.name})"
+
+    @property
+    @override
     def ndim_params(self):
         return self.base.ndim_params
 
     # pylint: disable=arguments-renamed, invalid-overridden-method
     @property
+    @override
     def has_sparse_matrix(self) -> bool:
         return self.base.has_sparse_matrix and isinstance(self.z, int)
 
     @staticmethod
+    @override
     def compute_matrix(base, z):  # pylint: disable=arguments-differ
         mat = qp.matrix(base)
         if isinstance(z, int):
@@ -123,6 +131,7 @@ class Pow2(SymbolicOp2):
 
     # pylint: disable=arguments-differ
     @staticmethod
+    @override
     def compute_sparse_matrix(base=None, z=0, format="csr"):
         if isinstance(z, int):
             base_matrix = base.compute_sparse_matrix(**base.arguments)
@@ -131,6 +140,7 @@ class Pow2(SymbolicOp2):
 
     # pylint: disable=arguments-renamed, invalid-overridden-method
     @property
+    @override
     def has_decomposition(self):
 
         if isinstance(self.z, int) and self.z > 0:
@@ -147,6 +157,7 @@ class Pow2(SymbolicOp2):
         return True
 
     @staticmethod
+    @override
     def compute_decomposition(base, z):
         try:
             return base.pow(z)
@@ -160,10 +171,12 @@ class Pow2(SymbolicOp2):
             raise DecompositionUndefinedError from e
 
     @property
+    @override
     def has_diagonalizing_gates(self):
         return self.base.has_diagonalizing_gates
 
     @staticmethod
+    @override
     def compute_diagonalizing_gates(base, z):  # pylint: disable=unused-argument
         r"""Sequence of gates that diagonalize the operator in the computational basis.
 
@@ -197,15 +210,18 @@ class Pow2(SymbolicOp2):
         return base.diagonalizing_gates()
 
     @staticmethod
+    @override
     def compute_eigvals(base, z):
         base_eigvals = base.eigvals()
         return [value**z for value in base_eigvals]
 
     # pylint: disable=arguments-renamed, invalid-overridden-method
     @property
+    @override
     def has_generator(self):
         return self.base.has_generator
 
+    @override
     def generator(self):
         r"""Generator of an operator that is in single-parameter-form.
 
@@ -220,14 +236,17 @@ class Pow2(SymbolicOp2):
         """
         return self.z * self.base.generator()
 
+    @override
     def pow(self, z):
         return [Pow2(base=self.base, z=self.z * z)]
 
     # pylint: disable=arguments-renamed, invalid-overridden-method
     @property
+    @override
     def has_adjoint(self):
         return isinstance(self.z, int)
 
+    @override
     def adjoint(self):
         """Create an operation that is the adjoint of this one.
 
@@ -255,6 +274,7 @@ class Pow2(SymbolicOp2):
             "The adjoint of Pow operators only is well-defined for integer powers."
         )
 
+    @override
     def simplify(self) -> Union["Pow", Identity]:
         # try using pauli_rep:
         if pr := self.pauli_rep:
